Better typing

+++ level1/source.c
@@ -1,11 +1,13 @@
 #include <stdio.h>
+#include <stdint.h>
+#include <stdlib.h>
 
 void run(void) {
   fwrite("Good... Wait what?\n", 1, 19, stdout);
   system("/bin/sh");
 }
 
-int main(void) {
+uint8_t main(void) {
   char buffer[64];
 
   gets(buffer);
